完成在线更新逻辑；

+++ Assets/Scripts/Model/Version.cs
@@ -9,7 +9,7 @@ public class ABVObject
     /// <summary>
     /// AB包大小
     /// </summary>
-    public long size;
+    public int size;
     /// <summary>
     /// AB包Hash值
     /// </summary>
